refactor: реорганизация структуры пакетов для устранения антипаттерна utils
- Перемещены валидаторы из 'utils.validators' в доменные пакеты:
  - FilmValidator → service.film.validation
  - UserValidator → service.user.validation
- Реструктурирован сервисный слой с доменной организацией:
  - Пакет service.film для компонентов фильмов
  - Пакет service.user для компонентов пользователей
- Обновлены импорты во всех зависимых классах
- Package utils.validators удален
