+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -4,7 +4,7 @@
  * Предоставляет REST API для работы с сущностью User.
  *
  * @see User
- * @see ru.yandex.practicum.filmorate.service.UserService
+ * @see ru.yandex.practicum.filmorate.service.user.UserService
  */
 package ru.yandex.practicum.filmorate.controller;
 
@@ -14,7 +14,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.service.UserService;
+import ru.yandex.practicum.filmorate.service.user.UserService;
 
 import java.util.List;
