+++ src/main/java/ru/yandex/practicum/filmorate/service/film/FilmService.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.service;
+package ru.yandex.practicum.filmorate.service.film;
 
 import ru.yandex.practicum.filmorate.model.Film;
