+++ src/test/java/ru/yandex/practicum/filmorate/service/FilmServiceImplTest.java
@@ -12,7 +12,8 @@ import ru.yandex.practicum.filmorate.exception.DuplicateException;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.managment.FilmStorage;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.utils.validators.FilmValidatorImpl;
+import ru.yandex.practicum.filmorate.service.film.FilmServiceImpl;
+import ru.yandex.practicum.filmorate.service.film.validation.FilmValidatorImpl;
 
 import java.time.LocalDate;
 import java.util.List;
